refactor(base): move `CollapseMenuStory`
