--- packages/@sanity/base/src/components/collapseMenu/__workshop__/index.ts
@@ -1,10 +1,0 @@
-import {defineScope} from '@sanity/ui-workshop'
-import {lazy} from 'react'
-
-export default defineScope('base/components', 'components', [
-  {
-    name: 'collapse-menu',
-    title: 'CollapseMenu',
-    component: lazy(() => import('./CollapseMenuStory')),
-  },
-])
